# table: reduce allocs in NewDestination by a factor of 3(ipv4)-5(ipv6)