@ishidawataru 2017-05-19
Thanks for the PR!

> I made sure to maintain the old behavior for IPv4 RadixKey's of omitting the last byte, I'm sure it's expected but annotating here in case the current behavior is an oversight. I.E.:
IpToRadixkey(net.ParseIP("10.3.32.0").To4(), 32)
== IpToRadixkey(net.ParseIP("10.3.32.100").To4(), 32)

I couldn't reproduce the behavior.
Could you explain a little bit more how to reproduce the behavior?

I tried below

```go
package main

import (
        "fmt"
        "net"

        "github.com/osrg/gobgp/table"
)

func main() {
        fmt.Println(table.IpToRadixkey(net.ParseIP("10.3.32.0").To4(), 32))
        fmt.Println(table.IpToRadixkey(net.ParseIP("10.3.32.100").To4(), 32))
}
```

and get

```
vagrant@vagrant ~ $ go run test.go
00001010000000110010000000000000
00001010000000110010000001100100
```

---

What about simplifying like below?

```go
func byteToBase2Buf(byt byte, buf []byte) {
        i := len(buf)
        for i > 0 {
                i--
                buf[i] = byte('0' + byt&1)
                byt >>= 1
        }
}
```