@cstockton 2017-05-19
Thanks for the comments, the note about the radix key was a annotation I made in my notes when I started that should have been omitted (before I tested it) sorry.

I simplified the byteToBase2Buf func similar to how you described, I also added the test I was using to ensure the output was identical, some of this pull req overlaps with [1334](https://github.com/osrg/gobgp/pull/1334) so I updated that one, I'll close this.